Translated unit tests to ginkgo and gomega
