+++ go.mod
@@ -12,7 +12,6 @@ require (
 	github.com/openshift/api v0.0.0-20250108172834-78bd56dba39b
 	github.com/openshift/client-go v0.0.0-20250106104058-89709a455e2a
 	github.com/openshift/cluster-version-operator v1.0.1-0.20250106161255-e90705bb5457
-	github.com/stretchr/testify v1.10.0
 	go.uber.org/zap v1.27.0
 	k8s.io/api v0.32.0
 	k8s.io/apimachinery v0.32.0
@@ -64,7 +63,6 @@ require (
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/operator-framework/api v0.17.1 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
 	github.com/prometheus/client_golang v1.20.5 // indirect
 	github.com/prometheus/client_model v0.6.1 // indirect
 	github.com/prometheus/common v0.61.0 // indirect
