+++ internal/operator/status_test.go
@@ -2,8 +2,10 @@ package operator
 
 import (
 	"fmt"
-	"testing"
 	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
 
 	configv1 "github.com/openshift/api/config/v1"
 	fakeconfigclient "github.com/openshift/client-go/config/clientset/versioned/fake"
@@ -12,7 +14,6 @@ import (
 	"github.com/openshift/vertical-pod-autoscaler-operator/internal/util"
 	"github.com/openshift/vertical-pod-autoscaler-operator/test/helpers"
 	appsv1 "k8s.io/api/apps/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -142,239 +143,216 @@ var (
 	})
 )
 
-func TestCheckCheckVPARecommender(t *testing.T) {
-	testCases := []struct {
-		label        string
-		expectedBool bool
-		expectedErr  error
-		objects      []runtime.Object
-	}{
-		{
-			label:        "no vertical-pod-autoscaler",
-			expectedBool: true,
-			expectedErr:  nil,
-			objects:      []runtime.Object{},
-		},
-		{
-			label:        "no deployment",
-			expectedBool: false,
-			expectedErr:  nil,
-			objects: []runtime.Object{
-				verticalPodAutoscaler,
+var _ = Describe("Testing Status", func() {
+
+	Describe("Check VPA Recommender", func() {
+		testCases := []struct {
+			label        string
+			expectedBool bool
+			expectedErr  error
+			objects      []runtime.Object
+		}{
+			{
+				label:        "no vertical-pod-autoscaler",
+				expectedBool: true,
+				expectedErr:  nil,
+				objects:      []runtime.Object{},
 			},
-		},
-		{
-			label:        "deployment wrong version",
-			expectedBool: false,
-			expectedErr:  nil,
-			objects: []runtime.Object{
-				verticalPodAutoscaler,
-				deployment.WithReleaseVersion("vBAD").Object(),
+			{
+				label:        "no deployment",
+				expectedBool: false,
+				expectedErr:  nil,
+				objects: []runtime.Object{
+					verticalPodAutoscaler,
+				},
 			},
-		},
-		{
-			label:        "deployment not available",
-			expectedBool: false,
-			expectedErr:  nil,
-			objects: []runtime.Object{
-				verticalPodAutoscaler,
-				deployment.WithAvailableReplicas(0).Object(),
+			{
+				label:        "deployment wrong version",
+				expectedBool: false,
+				expectedErr:  nil,
+				objects: []runtime.Object{
+					verticalPodAutoscaler,
+					deployment.WithReleaseVersion("vBAD").Object(),
+				},
 			},
-		},
-		{
-			label:        "available and updated",
-			expectedBool: true,
-			expectedErr:  nil,
-			objects: []runtime.Object{
-				verticalPodAutoscaler,
-				deployment.Object(),
+			{
+				label:        "deployment not available",
+				expectedBool: false,
+				expectedErr:  nil,
+				objects: []runtime.Object{
+					verticalPodAutoscaler,
+					deployment.WithAvailableReplicas(0).Object(),
+				},
 			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.label, func(t *testing.T) {
-			reporter := &StatusReporter{
-				client:       fakeclient.NewFakeClient(tc.objects...),
-				configClient: fakeconfigclient.NewSimpleClientset(),
-				config:       &TestStatusReporterConfig,
-			}
-
-			ok, err := reporter.CheckVPARecommender()
-
-			if ok != tc.expectedBool {
-				t.Errorf("got %t, want %t", ok, tc.expectedBool)
-			}
-
-			if !equality.Semantic.DeepEqual(err, tc.expectedErr) {
-				t.Errorf("got %v, want %v", err, tc.expectedErr)
-			}
-		})
-	}
-}
-
-func TestStatusChanges(t *testing.T) {
-	testCases := []struct {
-		label      string
-		expected   []configv1.ClusterOperatorStatusCondition
-		transition func(*StatusReporter) error
-	}{
-		{
-			label:    "available",
-			expected: AvailableConditions,
-			transition: func(r *StatusReporter) error {
-				return r.available("AvailableReason", "available message")
+			{
+				label:        "available and updated",
+				expectedBool: true,
+				expectedErr:  nil,
+				objects: []runtime.Object{
+					verticalPodAutoscaler,
+					deployment.Object(),
+				},
 			},
-		},
-		{
-			label:    "progressing",
-			expected: ProgressingConditions,
-			transition: func(r *StatusReporter) error {
-				return r.progressing("ProgressingReason", "progressing message")
-			},
-		},
-		{
-			label:    "degraded",
-			expected: DegradedConditions,
-			transition: func(r *StatusReporter) error {
-				return r.degraded("DegradedReason", "degraded message")
-			},
-		},
-	}
+		}
 
-	for _, tc := range testCases {
-		t.Run(tc.label, func(t *testing.T) {
-			reporter := &StatusReporter{
-				client:       fakeclient.NewFakeClient(),
-				configClient: fakeconfigclient.NewSimpleClientset(),
-				config:       &TestStatusReporterConfig,
-			}
-
-			err := tc.transition(reporter)
-			if err != nil {
-				t.Errorf("error applying status: %v", err)
-			}
-
-			co, err := reporter.GetClusterOperator()
-			if err != nil {
-				t.Errorf("error getting ClusterOperator: %v", err)
-			}
-
-			for _, cond := range tc.expected {
-				ok := cvorm.IsOperatorStatusConditionPresentAndEqual(
-					co.Status.Conditions, cond.Type, cond.Status,
-				)
-
-				if !ok {
-					t.Errorf("wrong status for condition: %s", cond.Type)
-				}
-			}
-		})
-	}
-}
-
-func TestReportStatus(t *testing.T) {
-	testCases := []struct {
-		label         string
-		versionChange bool
-		expectedBool  bool
-		expectedErr   error
-		expectedConds []configv1.ClusterOperatorStatusCondition
-		clientObjs    []runtime.Object
-		configObjs    []runtime.Object
-	}{
-		{
-			label:         "deployment wrong version",
-			versionChange: true,
-			expectedBool:  false,
-			expectedErr:   nil,
-			expectedConds: ProgressingConditions,
-			clientObjs: []runtime.Object{
-				verticalPodAutoscaler,
-				deployment.WithReleaseVersion("vWRONG").Object(),
-			},
-		},
-		{
-			label:         "available and updated",
-			versionChange: true,
-			expectedBool:  true,
-			expectedErr:   nil,
-			expectedConds: AvailableConditions,
-			clientObjs: []runtime.Object{
-				verticalPodAutoscaler,
-				deployment.WithReleaseVersion(ReleaseVersion).Object(),
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.label, func(t *testing.T) {
-			reporter := &StatusReporter{
-				client:       fakeclient.NewFakeClient(tc.clientObjs...),
-				configClient: fakeconfigclient.NewSimpleClientset(tc.configObjs...),
-				config:       &TestStatusReporterConfig,
-			}
-
-			ok, err := reporter.ReportStatus()
-
-			if ok != tc.expectedBool {
-				t.Errorf("got %t, want %t", ok, tc.expectedBool)
-			}
-
-			if !equality.Semantic.DeepEqual(err, tc.expectedErr) {
-				t.Errorf("got %v, want %v", err, tc.expectedErr)
-			}
-
-			// Check that the ClusterOperator status is created.
-			co, err := reporter.GetClusterOperator()
-			if err != nil {
-				t.Errorf("error getting ClusterOperator: %v", err)
-			}
-
-			// Check that all conditions have the expected status.
-			for _, cond := range tc.expectedConds {
-				ok := cvorm.IsOperatorStatusConditionPresentAndEqual(
-					co.Status.Conditions, cond.Type, cond.Status,
-				)
-
-				if !ok {
-					t.Errorf("wrong status for condition: %s", cond.Type)
-				}
-			}
-
-			// Check the LastTransitionTime of the Progressing condition.
-			for _, v := range co.Status.Versions {
-				if v.Name != "operator" {
-					continue
+		for _, tc := range testCases {
+			It(fmt.Sprintf("should have correct values when: %s", tc.label), func() {
+				reporter := &StatusReporter{
+					client:       fakeclient.NewFakeClient(tc.objects...),
+					configClient: fakeconfigclient.NewSimpleClientset(),
+					config:       &TestStatusReporterConfig,
 				}
 
-				p := cvorm.FindOperatorStatusCondition(
-					co.Status.Conditions, configv1.OperatorProgressing,
-				)
+				ok, err := reporter.CheckVPARecommender()
 
-				if p == nil {
-					t.Fatal("expected Progressing condition not found")
+				Expect(ok).To(Equal(tc.expectedBool), "got %t, want %t", ok, tc.expectedBool)
+				Expect(err).To(Equal(tc.expectedErr), "got %v, want %v", err, tc.expectedErr)
+			})
+		}
+	})
+
+	Describe("Testing Status Changes", func() {
+
+		testCases := []struct {
+			label      string
+			expected   []configv1.ClusterOperatorStatusCondition
+			transition func(*StatusReporter) error
+		}{
+			{
+				label:    "available",
+				expected: AvailableConditions,
+				transition: func(r *StatusReporter) error {
+					return r.available("AvailableReason", "available message")
+				},
+			},
+			{
+				label:    "progressing",
+				expected: ProgressingConditions,
+				transition: func(r *StatusReporter) error {
+					return r.progressing("ProgressingReason", "progressing message")
+				},
+			},
+			{
+				label:    "degraded",
+				expected: DegradedConditions,
+				transition: func(r *StatusReporter) error {
+					return r.degraded("DegradedReason", "degraded message")
+				},
+			},
+		}
+
+		for _, tc := range testCases {
+			It(fmt.Sprintf("should have correct status when operator is reporting as: %s", tc.label), func() {
+				reporter := &StatusReporter{
+					client:       fakeclient.NewFakeClient(),
+					configClient: fakeconfigclient.NewSimpleClientset(),
+					config:       &TestStatusReporterConfig,
 				}
 
-				switch tc.versionChange {
-				case true:
-					// If the version changed, the last transition time should
-					// be more recent than the original.
-					if !ConditionTransitionTime.Before(&p.LastTransitionTime) {
-						t.Error("expected Progressing condition transition time update")
+				err := tc.transition(reporter)
+				Expect(err).NotTo(HaveOccurred())
+
+				co, err := reporter.GetClusterOperator()
+				Expect(err).NotTo(HaveOccurred())
+
+				for _, cond := range tc.expected {
+					ok := cvorm.IsOperatorStatusConditionPresentAndEqual(
+						co.Status.Conditions, cond.Type, cond.Status,
+					)
+
+					Expect(ok).To(BeTrue(), "wrong status for condition: %s", cond.Type)
+				}
+			})
+		}
+	})
+
+	Describe("Test Report Status", func() {
+
+		testCases := []struct {
+			label         string
+			versionChange bool
+			expectedBool  bool
+			expectedErr   error
+			expectedConds []configv1.ClusterOperatorStatusCondition
+			clientObjs    []runtime.Object
+			configObjs    []runtime.Object
+		}{
+			{
+				label:         "deployment wrong version",
+				versionChange: true,
+				expectedBool:  false,
+				expectedErr:   nil,
+				expectedConds: ProgressingConditions,
+				clientObjs: []runtime.Object{
+					verticalPodAutoscaler,
+					deployment.WithReleaseVersion("vWRONG").Object(),
+				},
+			},
+			{
+				label:         "available and updated",
+				versionChange: true,
+				expectedBool:  true,
+				expectedErr:   nil,
+				expectedConds: AvailableConditions,
+				clientObjs: []runtime.Object{
+					verticalPodAutoscaler,
+					deployment.WithReleaseVersion(ReleaseVersion).Object(),
+				},
+			},
+		}
+
+		for _, tc := range testCases {
+			It(fmt.Sprintf("should report the status correctly when: %s", tc.label), func() {
+				reporter := &StatusReporter{
+					client:       fakeclient.NewFakeClient(tc.clientObjs...),
+					configClient: fakeconfigclient.NewSimpleClientset(tc.configObjs...),
+					config:       &TestStatusReporterConfig,
+				}
+
+				ok, err := reporter.ReportStatus()
+				Expect(ok).To(Equal(tc.expectedBool))
+				Expect(err).To(Equal(tc.expectedErr))
+
+				// Check that the ClusterOperator status is created.
+				co, err := reporter.GetClusterOperator()
+				Expect(err).NotTo(HaveOccurred())
+
+				// Check that all conditions have the expected status.
+				for _, cond := range tc.expectedConds {
+					ok := cvorm.IsOperatorStatusConditionPresentAndEqual(
+						co.Status.Conditions, cond.Type, cond.Status,
+					)
+					Expect(ok).To(BeTrue())
+				}
+
+				// Check the LastTransitionTime of the Progressing condition.
+				for _, v := range co.Status.Versions {
+					if v.Name != "operator" {
+						continue
 					}
 
-				case false:
-					// If the version did not change, the last transition time
-					// should remain unchanged.
-					if !ConditionTransitionTime.Equal(&p.LastTransitionTime) {
-						t.Error("unexpected Progressing condition transition time update")
-					}
+					p := cvorm.FindOperatorStatusCondition(
+						co.Status.Conditions, configv1.OperatorProgressing,
+					)
 
-				default:
-					panic("back away slowly...")
+					Expect(p).NotTo(BeNil())
+
+					switch tc.versionChange {
+					case true:
+						// If the version changed, the last transition time should
+						// be more recent than the original.
+						Expect(p.LastTransitionTime.Time).To(BeTemporally(">", ConditionTransitionTime.Time))
+
+					case false:
+						// If the version did not change, the last transition time
+						// should remain unchanged.
+						Expect(p.LastTransitionTime.Time).To(Equal(ConditionTransitionTime.Time))
+
+					default:
+						Fail("Unhandled test case; should not occur")
+					}
 				}
-			}
-		})
-	}
-}
+			})
+		}
+	})
+})
